refactor: renamed jsx files to tsx
